# Clean up as well as speed up the ILP extractor.