@mwillsey 2023-11-02
I agree, the ILP extractor needs a lot of work. Thanks for the PR.